chore: rename and update add sentence endpoint
