--- server/src/api/pending-sentences/index.ts
@@ -1,3 +1,0 @@
-import SentenceCollectorRouter from './routes'; 
-
-export default SentenceCollectorRouter;
