--- server/src/application/pending-sentences/use-case/command-handler/command/add-pending-sentence-command.ts
@@ -1,7 +1,0 @@
-export type AddPendingSentenceCommand = {
-  clientId: string
-  sentence: string
-  localeId: number
-  localeName: string
-  source: string
-}
