+++ server/src/application/sentences/use-case/command-handler/add-pending-sentence-vote-command-handler.ts
@@ -1,4 +1,4 @@
-import { insertPendingSentenceVoteIntoDb } from '../../repository/pending-sentences-repository'
+import { insertPendingSentenceVoteIntoDb } from '../../repository/sentences-repository'
 import { AddPendingSentenceVoteCommand } from './command/add-pending-sentence-vote-command'
 
 export const addPendingSentenceVoteCommandHandler = (
