+++ server/src/application/sentences/use-case/query-handler/get-pending-sentence-query-handler.ts
@@ -1,6 +1,6 @@
 import { pipe } from 'fp-ts/lib/function'
 import { task as T, taskOption as TO } from 'fp-ts'
-import { findPendingSentencesForReviewInDb } from '../../repository/pending-sentences-repository'
+import { findPendingSentencesForReviewInDb } from '../../repository/sentences-repository'
 import { GetPendingSentencesQuery } from './query/get-pending-sentences-query'
 import { PendingSentence } from '../../../../core/pending-sentences/types'
