+++ server/src/application/sentences/use-case/query-handler/query/get-pending-sentences-query.ts
