--- server/src/application/types/pending-sentence-submission.ts
@@ -1,6 +1,0 @@
-export type PendingSentenceSubmission = {
-  sentence: string;
-  source: string;
-  locale_id: number;
-  client_id: string;
-};
